Test app refactor
