+++ ts/packages/knowProTest/src/index.ts
@@ -1,5 +1,6 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-export * from "./knowproMemory.js";
-export * from "./searchLang.js";
+export * from "./knowproContext.js";
+export * from "./knowproSearch.js";
+export * from "./memoryCommands.js";
