--- ts/packages/knowProTest/src/knowproCommands.ts
@@ -1,2 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
