+++ ts/packages/knowProTest/src/knowproSearch.ts
@@ -2,35 +2,6 @@
 // Licensed under the MIT License.
 
 import * as kp from "knowpro";
-import * as cm from "conversation-memory";
-
-export async function getLangSearchResult(
-    conversation: kp.IConversation,
-    queryTranslator: kp.SearchQueryTranslator,
-    searchText: string,
-    options?: kp.LanguageSearchOptions,
-    langFilter?: kp.LanguageSearchFilter,
-    debugContext?: kp.LanguageSearchDebugContext,
-) {
-    const searchResults =
-        conversation instanceof cm.Memory
-            ? await conversation.searchWithLanguage(
-                  searchText,
-                  options,
-                  langFilter,
-                  debugContext,
-              )
-            : await kp.searchConversationWithLanguage(
-                  conversation,
-                  searchText,
-                  queryTranslator,
-                  options,
-                  langFilter,
-                  debugContext,
-              );
-
-    return searchResults;
-}
 
 export type LangSearchResults = {
     searchText: string;
